+++ ididi/__init__.py
@@ -10,17 +10,18 @@ Ididi is a pythonic dependency injection lib, with ergonomics api, without boilp
    >>> service = ididi.solve(MyService)
    >>> assert isinstance(service, MyService)
 
-is at <https://requests.readthedocs.io>.
+is at <https://github.com/raceychan/ididi>.
 
 :copyright: (c) 2024 by race chan.
 :license: MIT, see LICENSE for more details.
 """
 
-VERSION = "0.2.3"
+VERSION = "0.2.4"
 
 __version__ = VERSION
 
-from .api import entry as entry, solve as solve
+from .api import entry as entry
+from .api import solve as solve
 from .graph import DependencyGraph as DependencyGraph
 from .node import DependentNode as DependentNode
 from .visual import Visualizer as Visualizer